Remove PCAP code, using new PcapWriter
Also, given this is the first time running and capturing with
the newer writer code, fixed up a few oversight bugs, like, oh,
not actually _SETTING_ the file descriptor in the writer.

Plus. It became apparent that if the 'included_bytes' value
written to the file doesn't include the length of the Link
Layer header, then tools that read the file will have "garbled
data".