Add handling of SIGINT to allow proper cleanup.
This will assist with use of tools like Valgrind.

Also, A "two-step" handling of SIGINT takes place.
The callback to read packets could be waiting when the
'keep_looping' flag is changed. This means that the
main loop check may not happen until a packet is
received over the listened queue. If proper cleanup
isn't a concern or time (or patience) is short, a
second SIGINT will simply call exit(1).